worker/requirements.txt
python-nmap==0.7.1
redis==5.0.1
mac-vendor-lookup==0.1.12
requests==2.32.3
